+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapOptionsSink.java
@@ -55,6 +55,8 @@ interface GoogleMapOptionsSink {
 
   void setInitialCircles(Object initialCircles);
 
+  void setInitialHeatmaps(Object initialHeatmaps);
+
   void setInitialTileOverlays(List<Map<String, ?>> initialTileOverlays);
 
   void setMapStyle(@Nullable String style);