+++ packages/google_maps_flutter/google_maps_flutter_web/example/integration_test/google_maps_controller_test.dart
@@ -22,6 +22,7 @@ const String _kCloudMapId = '000000000000000'; // Dummy map ID.
 
 @GenerateNiceMocks(<MockSpec<dynamic>>[
   MockSpec<CirclesController>(),
+  MockSpec<HeatmapsController>(),
   MockSpec<PolygonsController>(),
   MockSpec<PolylinesController>(),
   MockSpec<MarkersController>(),
@@ -149,6 +150,20 @@ void main() {
           }, throwsAssertionError);
         });
 
+        testWidgets('cannot updateHeatmaps after dispose',
+            (WidgetTester tester) async {
+          controller.dispose();
+
+          expect(() {
+            controller.updateHeatmaps(
+              HeatmapUpdates.from(
+                const <Heatmap>{},
+                const <Heatmap>{},
+              ),
+            );
+          }, throwsAssertionError);
+        });
+
         testWidgets('cannot updatePolygons after dispose',
             (WidgetTester tester) async {
           controller.dispose();
@@ -219,6 +234,7 @@ void main() {
 
     group('init', () {
       late MockCirclesController circles;
+      late MockHeatmapsController heatmaps;
       late MockMarkersController markers;
       late MockPolygonsController polygons;
       late MockPolylinesController polylines;
@@ -227,6 +243,7 @@ void main() {
 
       setUp(() {
         circles = MockCirclesController();
+        heatmaps = MockHeatmapsController();
         markers = MockMarkersController();
         polygons = MockPolygonsController();
         polylines = MockPolylinesController();
@@ -239,6 +256,7 @@ void main() {
           ..debugSetOverrides(
             createMap: (_, __) => map,
             circles: circles,
+            heatmaps: heatmaps,
             markers: markers,
             polygons: polygons,
             polylines: polylines,
@@ -277,6 +295,7 @@ void main() {
           ..debugSetOverrides(
             createMap: (_, __) => map,
             circles: circles,
+            heatmaps: heatmaps,
             markers: markers,
             polygons: polygons,
             polylines: polylines,
@@ -285,6 +304,7 @@ void main() {
           ..init();
 
         verify(circles.bindToMap(mapId, map));
+        verify(heatmaps.bindToMap(mapId, map));
         verify(markers.bindToMap(mapId, map));
         verify(polygons.bindToMap(mapId, map));
         verify(polylines.bindToMap(mapId, map));
@@ -296,6 +316,17 @@ void main() {
           const Circle(
             circleId: CircleId('circle-1'),
             zIndex: 1234,
+          ),
+        }, heatmaps: <Heatmap>{
+          const Heatmap(
+            heatmapId: HeatmapId('heatmap-1'),
+            data: <WeightedLatLng>[
+              WeightedLatLng(LatLng(43.355114, -5.851333)),
+              WeightedLatLng(LatLng(43.354797, -5.851860)),
+              WeightedLatLng(LatLng(43.354469, -5.851318)),
+              WeightedLatLng(LatLng(43.354762, -5.850824)),
+            ],
+            radius: HeatmapRadius.fromPlatformSpecificValue(20),
           ),
         }, markers: <Marker>{
           const Marker(
@@ -342,6 +373,7 @@ void main() {
         controller = createController(mapObjects: mapObjects)
           ..debugSetOverrides(
             circles: circles,
+            heatmaps: heatmaps,
             markers: markers,
             polygons: polygons,
             polylines: polylines,
@@ -350,6 +382,7 @@ void main() {
           ..init();
 
         verify(circles.addCircles(mapObjects.circles));
+        verify(heatmaps.addHeatmaps(mapObjects.heatmaps));
         verify(markers.addMarkers(mapObjects.markers));
         verify(polygons.addPolygons(mapObjects.polygons));
         verify(polylines.addPolylines(mapObjects.polylines));
@@ -657,6 +690,53 @@ void main() {
         }));
         verify(mock.changeCircles(<Circle>{
           const Circle(circleId: CircleId('to-be-updated'), visible: false),
+        }));
+      });
+
+      testWidgets('updateHeatmaps', (WidgetTester tester) async {
+        final MockHeatmapsController mock = MockHeatmapsController();
+        controller.debugSetOverrides(heatmaps: mock);
+
+        final Set<Heatmap> previous = <Heatmap>{
+          const Heatmap(
+            heatmapId: HeatmapId('to-be-updated'),
+            radius: HeatmapRadius.fromPlatformSpecificValue(20),
+          ),
+          const Heatmap(
+            heatmapId: HeatmapId('to-be-removed'),
+            radius: HeatmapRadius.fromPlatformSpecificValue(20),
+          ),
+        };
+
+        final Set<Heatmap> current = <Heatmap>{
+          const Heatmap(
+            heatmapId: HeatmapId('to-be-updated'),
+            dissipating: false,
+            radius: HeatmapRadius.fromPlatformSpecificValue(20),
+          ),
+          const Heatmap(
+            heatmapId: HeatmapId('to-be-added'),
+            radius: HeatmapRadius.fromPlatformSpecificValue(20),
+          ),
+        };
+
+        controller.updateHeatmaps(HeatmapUpdates.from(previous, current));
+
+        verify(mock.removeHeatmaps(<HeatmapId>{
+          const HeatmapId('to-be-removed'),
+        }));
+        verify(mock.addHeatmaps(<Heatmap>{
+          const Heatmap(
+            heatmapId: HeatmapId('to-be-added'),
+            radius: HeatmapRadius.fromPlatformSpecificValue(20),
+          ),
+        }));
+        verify(mock.changeHeatmaps(<Heatmap>{
+          const Heatmap(
+            heatmapId: HeatmapId('to-be-updated'),
+            dissipating: false,
+            radius: HeatmapRadius.fromPlatformSpecificValue(20),
+          ),
         }));
       });