+++ packages/google_maps_flutter/google_maps_flutter_web/example/integration_test/google_maps_plugin_test.mocks.dart
@@ -136,6 +136,7 @@ class MockGoogleMapController extends _i1.Mock
     _i4.DebugSetOptionsFunction? setOptions,
     _i4.MarkersController? markers,
     _i4.CirclesController? circles,
+    _i4.HeatmapsController? heatmaps,
     _i4.PolygonsController? polygons,
     _i4.PolylinesController? polylines,
     _i4.TileOverlaysController? tileOverlays,
@@ -149,6 +150,7 @@ class MockGoogleMapController extends _i1.Mock
             #setOptions: setOptions,
             #markers: markers,
             #circles: circles,
+            #heatmaps: heatmaps,
             #polygons: polygons,
             #polylines: polylines,
             #tileOverlays: tileOverlays,
@@ -286,6 +288,14 @@ class MockGoogleMapController extends _i1.Mock
         returnValueForMissingStub: null,
       );
 
+  @override
+  void updateHeatmaps(_i2.HeatmapUpdates? updates) => super.noSuchMethod(
+        Invocation.method(
+          #updateHeatmaps,
+          [updates],
+        ),
+        returnValueForMissingStub: null,
+      );
   @override
   void updatePolygons(_i2.PolygonUpdates? updates) => super.noSuchMethod(
         Invocation.method(