+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.7.0
+
+* Adds support for heatmap layers.
+
 ## 2.6.0
 
 * Adds support for marker clustering.